+++ packages/view-main/src/shared/ipc/app/remote.ts
@@ -1,6 +1,7 @@
 import { showMessageBox } from '@/components/apis/dialog/messageBox'
 import { showInputBox } from '@/components/apis/inputModal/inputBox'
 import { showNotifyBox } from '@/components/apis/notify'
+import { setFullScreen } from '@/modules/app/store/action'
 import { extI18n } from '@/modules/extension/i18n'
 import { extensionState } from '@/modules/extension/store/state'
 
@@ -17,6 +18,9 @@ export default {
   },
   async winShow(show) {
     winShowEvent.emit(show)
+  },
+  async fullscreen(isFullscreen) {
+    setFullScreen(isFullscreen, true)
   },
   async showMessageBox(key, extId, options) {
     if (options.modal) {